ENH: Update to NetCDF v.3.6.2 (cvs merge -j netcdf-release) COMP: Set VTK_INSTALL_BIN_DIR_CM24 and VTK_INSTALL_LIB_DIR_CM24 in       CMakeLists.txt when VTK_INSTALL_NO_LIBRARIES is FALSE so that       build of vtknetcdf as a separate project (outside of VTK)       will work. BUG: Add COPYRIGHT,README,RELEASE_NOTES from NetCDF distribution      parent directory ENH: Add README describing what's in vtknetcdf as opposed to the      canonical NetCDF distro.

+++ VTK/Utilities/vtknetcdf/dim.c
@@ -2,7 +2,7 @@
  *  Copyright 1996, University Corporation for Atmospheric Research
  *      See netcdf/COPYRIGHT file for copying and redistribution conditions.
  */
-/* $Id: dim.c,v 1.4 2006-11-29 20:10:53 dcthomp Exp $ */
+/* $Id: dim.c,v 1.5 2007-08-27 20:54:06 dcthomp Exp $ */
 
 #include "nc.h"
 #include <stdlib.h>
@@ -99,7 +99,7 @@ find_NC_Udim(const NC_dimarray *ncap, NC_dim **dimpp)
   {
     /*EMPTY*/
   }
-  if(dimid >= (int)ncap->nelems)
+  if((size_t)dimid >= ncap->nelems)
     return(-1); /* not found */
   /* else, normal return */
   if(dimpp != NULL)
